mfd: twl: fix TWL6032 phy vbus detection
TWL6032 has a few charging registers prepended before the charging
registers the TWL6030 has. To be able to use common register defines
declare the additional registers as additional module.
At the moment this affects the access to CHARGERUSB_CTRL1 in phy-twl6030-usb.
Without this patch, it is accessing the wrong register on TWL6032.
The consequence is that presence of Vbus is not reported.

Signed-off-by: Andreas Kemnade <andreas@kemnade.info>